Improve Kkdf2MdlxParser cohesion a little bit.
